Update Portfolio component to use TypeScript and diversify projects
Added Ux Designer Portfolio piece to portfolio section. The portfolio component is now a TypeScript file to align with ongoing conversion to TypeScript across the project. Additionally, there is now a wider variety of projects being displayed instead of multiple instances of the same project, and the method of handling images has been simplified for improved efficiency.